Refactor HF clients jobs unit tests (#1213)
* Remove unused DEFAULT_SUMMARIZER_PROMPT

* hf_pipeline to use task_definition instead of task

* hf_pipeline to use task_definition instead of task

* Basic support for t5 family of models

* Fix to include lumigator schemas in jobs

* Directly import tasks from lumigator_schemas thanks to ray py_modules settings

* Prefix based HF models: mt0 and flan-t5 families

* Dealing with some model configs not having any of the plausible_max_length_params; use default value

* Add langcodes to job requirements

* Handling language name/code in prefix and pipeline task for PrefixTranslationClient with langcodes

* Fix jobs unit test with updated fixture

* task_definition as InferenceJobConfig field

* Add HuggingFaceLanguageCodeTranslationClient

* Put back HuggingFaceLanguageCodeTranslationClient

* Revert to task

* Move LanguageCodesSetupMixin to separate file

* Remove duplication of set_seq2seq_max_length

* Fix jobs unit test

* Add test for TestLanguageCodesSetupMixin

* Move logic to load_translation_config

* Added unit test for HuggingFaceLanguageCodeTranslationClient

* Bug fix with args order

* Reusable components for the tests in MockBaseSetup

* Reorganize folders

* Split up test_huggingface_clients

* Fix imports

---------
